implementing security for endpoints

+++ src/main/java/by/ares/orderservice/controller/OrderItemController.java
@@ -2,6 +2,7 @@ package by.ares.orderservice.controller;
 
 import by.ares.orderservice.dto.response.OrderDto;
 import by.ares.orderservice.service.OrderItemService;
+import by.ares.orderservice.service.SecurityValidationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,10 +14,14 @@ import org.springframework.web.bind.annotation.*;
 public class OrderItemController {
 
     private final OrderItemService orderItemService;
+    private final SecurityValidationService securityValidationService;
 
     @PutMapping("/items/{itemId}")
     public ResponseEntity<OrderDto> addItem(@PathVariable Long orderId,
-                                             @PathVariable Long itemId) {
+                                            @PathVariable Long itemId,
+                                            @RequestHeader("X-User-Id") Long userId,
+                                            @RequestHeader("X-User-Role") String role) {
+        securityValidationService.validateOrderAccess(orderId, userId, role);
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(orderItemService.addItemToOrder(orderId, itemId));
@@ -24,7 +29,10 @@ public class OrderItemController {
 
     @DeleteMapping("/items/{itemId}")
     public ResponseEntity<OrderDto> removeItem(@PathVariable Long orderId,
-                                                @PathVariable Long itemId) {
+                                               @PathVariable Long itemId,
+                                               @RequestHeader("X-User-Id") Long userId,
+                                               @RequestHeader("X-User-Role") String role) {
+        securityValidationService.validateOrderAccess(orderId, userId, role);
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(orderItemService.removeItemFromOrder(orderId, itemId));
